requirements.txt
numpy==1.16.4
Pillow==6.1.0
torch==1.3.1
torchvision==0.4.2